Make emscripten to generate an AudioWorklet compatible file

+++ src/base.h
@@ -13,8 +13,8 @@ typedef unsigned short uint16_t;
 typedef __uint32_t uint32_t;
 
 #else
-typedef signed long int32_t;
-typedef unsigned long uint32_t;
+typedef signed int int32_t;
+typedef unsigned int uint32_t;
 #endif
 
 
@@ -23,9 +23,9 @@ typedef unsigned long uint32_t;
 //#define DEBUG
 
 /*
-* turn on trace output to debug ADSR handling of a specific voice: 
+* turn on trace output to debug ADSR handling of a specific voice:
 * debug output will be printed to console and audio output will be
-* limited to the specified voice and the specified frame-range 
+* limited to the specified voice and the specified frame-range
 */
 
 #define PSID_DEBUG_VOICE 1
@@ -40,14 +40,14 @@ typedef enum {
 	DigiNone = 0,
 	DigiD418 = 1,			// legacy 4-bit approach
 	DigiMahoneyD418 = 2,	// Mahoney's "8-bit" D418 samples
-	
+
 	// all the filterable digi types
 	DigiFM = 3,				// 8-bit frequency modulation; e.g. Vicious_SID_2-15638Hz.sid, LMan - Vortex.sid, etc
 	DigiPWM = 4,			// older PWM impls, e.g. Bla_Bla.sid, Bouncy_Balls_RCA_Intro.sid
 	DigiPWMTest = 5			// new test-bit based PWM; e.g. Wonderland_XII-Digi_part_1.sid, GhostOrGoblin.sid, etc
 } DigiType;
 
-#define MAX_SIDS 10			// might eventually need to be increased 
+#define MAX_SIDS 10			// might eventually need to be increased
 
 
 #endif
